extracting api responses to their own files to improve ide performance
