+++ matrix/services/sync/src/main/kotlin/app/dapk/st/matrix/sync/internal/request/EncryptedContentDeserializer.kt
@@ -1,0 +1,29 @@
+package app.dapk.st.matrix.sync.internal.request
+
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.descriptors.buildClassSerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
+import kotlinx.serialization.json.JsonDecoder
+import kotlinx.serialization.json.jsonObject
+import kotlinx.serialization.json.jsonPrimitive
+
+internal object EncryptedContentDeserializer : KSerializer<ApiEncryptedContent> {
+
+    override val descriptor: SerialDescriptor = buildClassSerialDescriptor("encryptedContent")
+
+    override fun deserialize(decoder: Decoder): ApiEncryptedContent {
+        require(decoder is JsonDecoder)
+        val element = decoder.decodeJsonElement()
+        return when (val algorithm = element.jsonObject["algorithm"]?.jsonPrimitive?.content) {
+            "m.olm.v1.curve25519-aes-sha2" -> ApiEncryptedContent.OlmV1.serializer().deserialize(decoder)
+            "m.megolm.v1.aes-sha2" -> ApiEncryptedContent.MegOlmV1.serializer().deserialize(decoder)
+            null -> ApiEncryptedContent.Unknown
+            else -> throw IllegalArgumentException("Unknown algorithm : $algorithm")
+        }
+    }
+
+    override fun serialize(encoder: Encoder, value: ApiEncryptedContent) = TODO("Not yet implemented")
+
+}
